Reorganize legacy code structure for Phase 1 preparation
- Move workspace_manager and document_manager to services/legacy/
- Update all include paths to new legacy locations
- Reorganize test structure to match code organization
- Update BUILD.bazel file paths for moved tests
- Consolidate all legacy-specific code for easier future removal
