+++ include/slangd/services/legacy/definition_provider.hpp
@@ -8,8 +8,8 @@
 #include <spdlog/spdlog.h>
 
 #include "lsp/basic.hpp"
-#include "slangd/core/document_manager.hpp"
-#include "slangd/core/workspace_manager.hpp"
+#include "slangd/services/legacy/document_manager.hpp"
+#include "slangd/services/legacy/workspace_manager.hpp"
 #include "slangd/services/legacy/language_feature_provider.hpp"
 
 namespace slangd {
