+++ include/slangd/services/legacy/diagnostics_provider.hpp
@@ -11,8 +11,8 @@
 #include <slang/syntax/SyntaxTree.h>
 #include <slang/text/SourceManager.h>
 
-#include "slangd/core/document_manager.hpp"
-#include "slangd/core/workspace_manager.hpp"
+#include "slangd/services/legacy/document_manager.hpp"
+#include "slangd/services/legacy/workspace_manager.hpp"
 #include "slangd/services/legacy/language_feature_provider.hpp"
 
 namespace slangd {
