+++ include/slangd/services/legacy/document_manager.hpp
