+++ include/slangd/services/legacy/workspace_manager.hpp
