+++ test/slangd/services/legacy/definition_provider_test.cpp
