+++ test/slangd/services/legacy/diagnostics_provider_test.cpp
