+++ test/slangd/services/legacy/symbols_provider_test.cpp
